V-0.0.38

+++ src/lib/bin/cli.ts
@@ -24,14 +24,17 @@ const printUsageDetails = () => {
 };
 const svelteWayDev = async () => {
 	try {
-		const currentDirectory = cwd().replace(/\\/g, '/')
+		const currentDirectory = cwd().replace(/\\/g, '/');
 		if (fs.existsSync(`${currentDirectory}/Svelteway-safe-layout.svelte`)) {
-			const content = await fs.promises.readFile(`${currentDirectory}/Svelteway-safe-layout.svelte`);
-			const buf = Buffer.from(`${content}`, 'utf8')
+			const content = await fs.promises.readFile(
+				`${currentDirectory}/Svelteway-safe-layout.svelte`
+			);
+			const buf = Buffer.from(`${content}`, 'utf8');
 			fs.writeFileSync(`${currentDirectory}/src/routes/+layout.svelte`, buf);
 		}
 		const folderName = `${currentDirectory}/src/routes/api/svelteway`;
-		const buf = Buffer.from(`import type { Actions } from './$types';
+		const buf = Buffer.from(
+			`import type { Actions } from './$types';
 		import { redirect } from '@sveltejs/kit';
 		import { Buffer } from 'node:buffer';
 		import * as fs from 'fs';
@@ -51,15 +54,21 @@ const svelteWayDev = async () => {
 					console.error(err);
 				}
 			}
-		} satisfies Actions;`, 'utf8');
+		} satisfies Actions;`,
+			'utf8'
+		);
 		try {
 			if (!fs.existsSync(folderName)) {
-			  fs.mkdirSync(folderName)
-			  fs.writeFileSync(`${folderName}/+page.server.ts`, buf);
+				fs.mkdirSync(folderName);
+				if (fs.existsSync(`${currentDirectory}/Svelteway-safe-layout.svelte`)) {
+				fs.writeFileSync(`${folderName}/+page.server.ts`, buf);
+				} else {
+					fs.writeFileSync(`${folderName}/+page.server.js`, buf);
+				}
 			}
-		  } catch (err) {
+		} catch (err) {
 			console.error(err);
-		  }
+		}
 	} catch (err) {
 		console.error(err);
 	}
@@ -67,32 +76,32 @@ const svelteWayDev = async () => {
 };
 const svelteWayBuild = async () => {
 	try {
-		const currentDirectory = cwd().replace(/\\/g, '/')
-		const layoutFile = `${currentDirectory}/src/routes/+layout.svelte`
-  		const content = await fs.promises.readFile(layoutFile)
-		const buf = Buffer.from(`${content}`, 'utf8')
+		const currentDirectory = cwd().replace(/\\/g, '/');
+		const layoutFile = `${currentDirectory}/src/routes/+layout.svelte`;
+		const content = await fs.promises.readFile(layoutFile);
+		const buf = Buffer.from(`${content}`, 'utf8');
 		fs.writeFileSync(`${currentDirectory}/Svelteway-safe-layout.svelte`, buf);
-		fs.rm(`${currentDirectory}/src/routes/api/svelteway`, { recursive: true });
+		if (fs.existsSync(`${currentDirectory}/src/routes/api/svelteway`)) {
+			fs.rm(
+				`${currentDirectory}/src/routes/api/svelteway`,
+				{ recursive: true, force: true },
+				(err) => {
+					return;
+				}
+			);
+		}
 		const options = {
 			files: layoutFile,
-			from: '<Layout data={data}>',
-			to: '',
-		  };
-		const options2 = {
-			files: layoutFile,
-			from: '</Layout>',
-			to: '',
-		  };
-		const options3 = {
-			files: layoutFile,
-			from: `import Layout from '$lib/components/Layout.svelte';`,
-			to: '',
-		  };
-		  const results = await replaceInFile(options)
-		  const results2 = await replaceInFile(options2)
-		  const results3 = await replaceInFile(options3)
+			from: [
+				/<Layout \{data\}>/g,
+				/<\/Layout>/g,
+				/import Layout from '\$lib\/components\/Layout\.svelte';/g
+			],
+			to: ''
+		};
+		const results = await replaceInFile(options);
 	} catch (err) {
-		console.error(err);
+		return;
 	}
 	process.exit(0);
 };
